Finishing touches
Not as polished as it really could be but I think this is worth
submitting now.

Added score uploading (getting the key from a not-committed header
file), main menu, and some fun facts to the game-over screen!

Special thanks to CERN for having nothing at all to do with this game
and also doing wicked science stuff.