Huge changes/re-organizing the whole project
Removed the Ludum Hadron Collider game code.

Switched to a single Makefile for building. It has the targets
"debug_linux" (default, for development), as well as "release_linux" and
"release_win32" to build (or cross-build) release binaries which have
the shader sources compiled in statically and don't have debug symbols.

Added a pg_game_state structure which forms a convenient abstraction
layer on top of any game state, and handles the time-step for it.
Included a very simple example game state as a demonstration/template.

Separated libraries into linux and windows sections, for static linking
for either target platform. (Currently only libcurl is linked statically
on Linux, while on Windows I static link with libcurl, SDL2, and GLEW.)