Slightly changing the cylinder, cone, and truncated cone primitives
Now the faces are aligned on the angle steps, rather than the vertices. I hope
this will make modifying/extruding faces easier because now (for models
generated with the right number of segments) an extrusion can still be aligned
on an imaginary grid rather than having to extend at some strange angle.