Preliminary implementation of signed distance function operation tree
There's a lot here, but it's basically just a binary tree implementation of
signed distance function modeling. Each node is an operation or a primitive,
they combine recursively in a binary tree to generate signed distances to an
implicit surface, which can then be used for other purposes like marching cubes,
ray-marching, or other things. Also included is a function for parsing textual
representations of SDF trees like so:

(intersect (BOX 1 0.5 0.5) (union (CYLINDER 2 0.1) (TORUS 0.8 0.1)))

This syntax will hopefully be documented. Eventually.

It also uses the tool 'gperf' to generate a perfect hash function for the SDF
text parsing keywords. The makefile invokes it to produce the file
`sdf_keywords.gperf.c` which is then included directly by `src/procgl/sdf.c`.

Next is a marching cubes implementation for generating models!