Huge overhaul of the model system
Rather than having several vertex types, now each model will keep a separate
array for each of its components, like position, uv, normals, etc. As well as
a bitmask with a bit for each component, to check if a model has a certain one.

Then, the model must be attached to a shader with "pg_shader_buffer_model" which
will construct a GL buffer with all the vertex definitions for the intersection
of components between the shader and the model.

So before a model's vertices were laid out like: (pos, uv, norm, tan,
bitan)(pos, uv, norm, tan, bitan)...

Now a model's vertices are laid out likeA: (pos, pos, pos...)(uv, uv,
uv...)(tan, tan, tan...) and so on.

So if a model has pos, uv, and color components, and a shader only uses vertex
colors, not texturing, then when "pg_shader_buffer_model" is called with the
two of them, a VBO will be generated with only the pos and uv components. If a
model has already been buffered for another shader with identical requirements,
then the buffers will be shared.

Also, the pg_shader structure is now aware of the basic pg_model components,
and they can be linked just like the matrices at shader creation. All together,
this should make things quite a bit simpler to use.

Also removing "shapes", they are just a shadow of models anyway, and it will be
better if 2d stuff can be implemented with the same model and shader
infrastructure as 3d.