Refactoring the game states, mainly the editor state
Now the editor has its own version of the map struct which is simpler and leaves
out gameplay data like position, velocity, etc. It writes this format to a file
and loads it from a file, then expands this format to the full map structure.

Also the map is back to being one single chunk rather than several seprate
areas.