Preliminary path-finding implementation
This works by performing a breadth-first search away from the player's currenet
cell, marking each one with a value that decrements with each step away from the
player, down to 0. Then, entities wanting to pathfind to the player will simply
move up this distance gradient.

The breadth-first search fills the level using a "walkability" criterion. The
"walkability" (and certain special flags for handling ramps and half-tiles) of
each tile is pre-calculated when the map is loaded.