Beginning a large overhaul of the shader and model handling system
This is meant to make the shader system a bit more generic, and also allow a
more proper core render loop.

I'd rather not have a bunch of "pg_shader_<name>_<action>" functions for every
single built-in shader, resulting in weird quirks and interface inconsistencies
between all of them (not to mention writing any new shader requires writing a
load of these wrapper functions). So this new system encapsulates shaders into a
system of pg_renderpass objects, which have a table of uniforms, options,
matrices. All of this is given to the shader in a generic way, so writing a
shader you only need to add a uniform with a name, and then add an entry to the
renderpass uniform table, and it will get sent to the shader each time the
renderpass is used in a frame. The pg_renderpass object then contains an array
of pg_rendergroup objects, which contain an unstructured array of uniform data,
and a function pointer to a function which consumes the array items in whatever
way is required for the shader, with a draw call being issued between each call
to the function pointer. The function pointer has access to whatever uniforms
are provided in a call to the function pg_rendergroup_require(), in the order
they are given, as well as the full set of common matrices for the current
renderpass which is being executed.